backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
requests==2.31.0
bcrypt==4.0.1